requirements.txt
aniso8601==8.0.0
aspy.yaml==1.3.0
certifi==2019.9.11
cfgv==2.0.1
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
identify==1.4.7
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
nodeenv==1.3.3
pre-commit==1.20.0
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
SQLAlchemy==1.3.11
toml==0.10.0
virtualenv==16.7.7
Werkzeug==0.16.0
zipp==0.6.0
